feat: remove duplicate lines
